pyproject.toml
[project]
name = "evidenceplatform"
version = "0.1.0"
description = "Evidence archiving and browsing platform"
requires-python = ">=3.12"
dependencies = [
    "altgraph==0.17.4",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "archiveis==0.0.9",
    "authlib==1.6.0",
    "backports-datetime-fromisoformat==2.0.3",
    "bcrypt==4.3.0",
    "beautifulsoup4==4.13.4",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "click==8.1.8",
    "colorama==0.4.6",
    "cryptography~=41.0.7",
    "dparse==0.6.4",
    "exceptiongroup==1.3.0",
    "filelock==3.16.1",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "har2warc==1.0.4",
    "http-status==1.0.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "ijson==3.4.0",
    "importlib-metadata==8.7.0",
    "jinja2==3.1.6",
    "joblib==1.5.1",
    "markdown-it-py==3.0.0",
    "markupsafe==3.0.2",
    "marshmallow==4.0.0",
    "mdurl==0.1.2",
    "mouseinfo==0.1.3",
    "mysql-connector-python==9.3.0",
    "nltk==3.9.1",
    "numpy==2.3.1",
    "packaging==25.0",
    "pefile==2023.2.7",
    "pillow==11.1.0",
    "playwright==1.51.0",
    "psutil==6.1.1",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pyautogui==0.9.54",
    "pycparser==2.22",
    "pydantic==2.9.2",
    "pydantic-core==2.23.4",
    "pyee==12.1.1",
    "pygetwindow==0.0.9",
    "pygments==2.19.2",
    "pyinstaller==6.14.0",
    "pyinstaller-hooks-contrib==2025.4",
    "pymsgbox==1.0.9",
    "pyperclip==1.9.0",
    "pyrect==0.2.0",
    "pyscreeze==1.0.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytweening==1.2.0",
    "pywin32-ctypes==0.2.3",
    "regex==2024.11.6",
    "requests==2.32.3",
    "rfc3161ng==2.1.3",
    "rich==14.0.0",
    "ruamel.yaml==0.18.14",
    "ruamel.yaml.clib==0.2.12",
    "safety==3.5.1",
    "safety-schemas==0.0.14",
    "shellingham==1.5.4",
    "six==1.17.0",
    "sniffio==1.3.1",
    "soupsieve==2.7",
    "tenacity==9.1.2",
    "tomli==2.2.1",
    "tomlkit==0.13.2",
    "tqdm==4.67.1",
    "typer==0.16.0",
    "typing-inspection==0.4.0",
    "typing-extensions==4.13.1",
    "urllib3==2.5.0",
    "warcio==1.7.5",
    "zipp==3.23.0",
    "zstandard==0.23.0",
    "fastapi>=0.123.0",
    "uvicorn>=0.38.0",
    "argon2-cffi>=25.1.0",
    "tzlocal>=5.3.1",
    "pytz>=2025.2",
    "opencv-python-headless>=4.11.0.86",
]

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
